+++ src/pixClaim/log.php
@@ -19,7 +19,7 @@ class Log extends Resource
     
     ## Attributes:
         - id [string]: unique id returned when the log is created. ex: "5656565656565656"
-        - created [DateTime or string, default null]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
+        - created [DateTime]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
         - type [string]: type of the PixClaim event which triggered the log creation. ex: "created" or "failed"
         - errors [list of strings]: list of errors linked to this PixClaim event
         - agent [string]: agent that modified the PixClaim resulting in the Log. Options: "claimer", "claimed".
@@ -36,6 +36,7 @@ class Log extends Resource
         $this-> type = Checks::checkParam($params, "type");
         $this-> errors = Checks::checkParam($params, "errors");
         $this-> agent = Checks::checkParam($params, "agent");
+        $this-> reason = Checks::checkParam($params, "reason");
         $this-> claim = Checks::checkParam($params, "claim");
 
         Checks::checkParams($params);
@@ -71,7 +72,7 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: ["created"] or ["failed"]
+        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "confirming", "confirmed", "success", "canceling" and "canceled".
         - claimIds [list of strings, default null]: list of PixClaim ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if starkinfra\user was set before function call
     
@@ -97,7 +98,7 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: ["created"] or ["failed"]
+        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "confirming", "confirmed", "success", "canceling" and "canceled".
         - claimIds [list of strings, default null]: list of PixClaim IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call