+++ src/pixInfraction/log.php
@@ -19,10 +19,10 @@ class Log extends Resource
     
     ## Attributes:
         - id [string]: unique id returned when the log is created. ex: "5656565656565656"
-        - created [DateTime or string, default null]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
-        - type [string]: type of the PixInfraction event which triggered the log creation. ex: "created", "registered", "updated", "failed", "canceling" and "canceled".
+        - created [Date, Datetime or string]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
+        - type [string]: type of the PixInfraction event which triggered the log creation. ex: "created", "failed", "delivering", "delivered", "closed", "canceled".
         - errors [list of strings]: list of errors linked to this PixInfraction event
-        - report [PixInfraction]: PixInfraction entity to which the log refers to.
+        - infraction [PixInfraction]: PixInfraction entity to which the log refers to.
      */
 
     function __construct(array $params)
@@ -32,7 +32,7 @@ class Log extends Resource
         $this-> created = Checks::checkDateTime(Checks::checkParam($params, "created"));
         $this-> type = Checks::checkParam($params, "type");
         $this-> errors = Checks::checkParam($params, "errors");
-        $this-> report = Checks::checkParam($params, "report");
+        $this-> infraction = Checks::checkParam($params, "infraction");
 
         Checks::checkParams($params);
     }
@@ -66,8 +66,8 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "success" or "failed"
-        - PixInfraction [list of strings, default null]: list of PixInfraction IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "closed", "canceled".
+        - infractionsIds [list of strings, default null]: list of PixInfraction IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if starkinfra\user was set before function call
     
     ## Return:
@@ -94,8 +94,8 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "success" or "failed"
-        - reportIds [list of strings, default null]: list of PixInfraction ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - types [list of strings, default null]: filter retrieved objects by types. ex:  "created", "failed", "delivering", "delivered", "closed", "canceled".
+        - infractionIds [list of strings, default null]: list of PixInfraction ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if starkinfra\user was set before function call
     
     ## Return: