+++ src/pixKey/log.php
@@ -20,7 +20,7 @@ class Log extends Resource
     
     ## Attributes:
         - id [string]: unique id returned when the log is created. ex: "5656565656565656"
-        - created [DateTime, default null]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
+        - created [Date, Datetime or string]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
         - type [string]: type of the PixKey event which triggered the log creation. ex: "created", "registered", "updated", "failed", "canceling" and "canceled".
         - errors [list of strings]: list of errors linked to this PixKey event
         - key [PixKey]: PixKey entity to which the log refers to.
@@ -68,7 +68,7 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: ["created"] or ["failed"]
+        - types [list of strings, default null]: filter retrieved objects by types. ex: "created","registered","updated","failed","canceling" and "canceled".
         - keyIds [list of strings, default null]: list of PixKey IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call
     
@@ -94,7 +94,7 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: ["created"] or ["failed"]
+        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "registered", "updated", "failed", "canceling" and "canceled".
         - keyIds [list of strings, default null]: list of PixKey IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call