+++ src/pixReversal/log.php
@@ -29,10 +29,10 @@ class Log extends Resource
     {
         parent::__construct($params);
 
-        $this-> created = Checks::checkDateTime(Checks::checkParam($params, "created"));
-        $this-> type = Checks::checkParam($params, "type");
-        $this-> errors = Checks::checkParam($params, "errors");
         $this-> reversal = Checks::checkParam($params, "reversal");
+        $this-> errors = Checks::checkParam($params, "errors");
+        $this-> type = Checks::checkParam($params, "type");
+        $this-> created = Checks::checkDateTime(Checks::checkParam($params, "created"));
 
         Checks::checkParams($params);
     }
@@ -65,7 +65,7 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [array of strings, default null]: filter retrieved objects by types. ex: "success" or "failed"
+        - types [array of strings, default null]: filter retrieved objects by types. ex: "sent", "denied", "failed", "created", "success", "approved", "credited", "refunded", "processing".
         - reversalIds [array of strings, default null]: array of PixReversal ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
     
@@ -90,7 +90,7 @@ class Log extends Resource
         - limit [integer, default 100]: maximum number of objects to be retrieved. It must be an integer between 1 and 100. ex: 50
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "success" or "failed"
+        - types [list of strings, default null]: filter retrieved objects by types. ex: "sent", "denied", "failed", "created", "success", "approved", "credited", "refunded", "processing".
         - reversalIds [list of strings, default null]: list of PixReversal ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call